DERIVATIVE INFO INCLUDED INTO NEWTON-RAPHSON
Code was throwing errors. Based on Leo's sugestions, I cleared the cache (~/.cache/qnm) and in the Jupyter notebook, I comment out the link which downloads the qnm data. The code now runs well and gives correct results. It also runs slowly since the use of cache is not made.