An error got introduced when branch-merge happened (commit: 610bd350). We fix it here. The code runs very slowly though.
What was that error: In nearby.py, I was feeding the function which was a 0th inversion of the continued fraction into the Newton-Raphson routine. And I was always getting the 0th overtone.

BDW, I don't understand why there should be an error. It should give the correct overtone even if I feed a different inversion of the continued fraction, provided the seed guess and derivative for the Newton-Raphson routine are correct. Try to understand it.